GeoGuesser/requirements.txt
matplotlib==3.8.1
pandas==2.1.2
Pillow==10.1.0
tensorflow==2.14.0
scikit-plot==0.3.7
opencv-python==4.8.1.78
plotly==5.18.0
